Split up program into separate files
